gcp-audio-processing/requirements.txt
demucs
openai-whisper
firebase-admin
flask
gunicorn
requests
